requirements.txt
# Core
pydantic>=2.0
python-dotenv>=1.0
httpx>=0.27
rich>=13.0
click>=8.1

# Google AI (Gemini)
google-genai>=1.0

# Anthropic
anthropic>=0.40

# OpenAI (optional)
openai>=1.50

# fal.ai generation API
fal-client>=0.5

# Video preprocessing
scenedetect[opencv]>=0.6
opencv-python-headless>=4.9

# Audio
openai-whisper>=20231117
demucs>=4.0
librosa>=0.10

# ML / embeddings
torch>=2.1
transformers>=4.40
open-clip-torch>=2.24
numpy>=1.26
Pillow>=10.0
